+++ lib/src/indicator/classic_indicator.dart
@@ -125,20 +125,18 @@ class _ClassicHeaderState extends RefreshIndicatorState<ClassicHeader> {
                     ? widget.failedIcon
                     : mode == RefreshStatus.canTwoLevel
                         ? widget.canTwoLevelIcon
-                        : mode == RefreshStatus.canTwoLevel
-                            ? widget.canTwoLevelIcon
-                            : mode == RefreshStatus.refreshing
-                                ? widget.refreshingIcon ??
-                                    SizedBox(
-                                      width: 25.0,
-                                      height: 25.0,
-                                      child: defaultTargetPlatform ==
-                                              TargetPlatform.iOS
-                                          ? const CupertinoActivityIndicator()
-                                          : const CircularProgressIndicator(
-                                              strokeWidth: 2.0),
-                                    )
-                                : widget.twoLevelView;
+                        : mode == RefreshStatus.refreshing
+                            ? widget.refreshingIcon ??
+                                SizedBox(
+                                  width: 25.0,
+                                  height: 25.0,
+                                  child: defaultTargetPlatform ==
+                                          TargetPlatform.iOS
+                                      ? const CupertinoActivityIndicator()
+                                      : const CircularProgressIndicator(
+                                          strokeWidth: 2.0),
+                                )
+                            : widget.twoLevelView;
     return icon ?? Container();
   }