+++ lib/src/indicator/custom_indicator.dart
@@ -154,12 +154,8 @@ class CustomFooter extends LoadIndicator {
     this.endLoading,
     LoadStyle loadStyle = LoadStyle.ShowAlways,
     required this.builder,
-    Function? onClick,
-  }) : super(
-            key: key,
-            onClick: onClick as void Function()?,
-            loadStyle: loadStyle,
-            height: height);
+    VoidCallback? onClick,
+  }) : super(key: key, onClick: onClick, loadStyle: loadStyle, height: height);
 
   @override
   State<StatefulWidget> createState() {